// src/fsd/3-features/teams/teams.enums.ts
export enum GameMode {
    tournamentArena = '_ta',
    guildRaids = '_gr',
    guildWar = '_gw',
    legendaryRelease = '_lre',
    incursion = '_inc',
    survival = '_sur',
}

export enum GuildRaidBoss {
    avatar = '_avatar',
    ghazghkull = '_ghaz',
    hiveTyrant = '_tyrant',
    magnus = '_magnus',
    mortarion = '_mortarion',
    rogalDorn = '_tank',
    screamerKiller = '_screamer',
    szarekh = '_szarekh',
    tervigon = '_tervigon',
    cawl = '_cawl',
}

export enum TaMode {
    conquest = '_con',
    powerUps = '_pu',
    draftPowerUps = '_dpu',
    infestedPowerUps = '_ipu',
}

export enum GwMode {
    offense = '_off',
    defense = '_def',
}